+++ tests/input_files/series/flowmap/01.js
@@ -1,6 +1,6 @@
 {
-  animationLimit: 10,
   colorAxis: 1,
+  animationLimit: 10,
   colorIndex: 5,
   colorKey: 'some-key-value',
   findNearestPointBy: 'x',